Unit Tests
Code to bring in unit testing for all methods that perform a function with exception to ApplicationService as that simple runs the other methods. Some code has also been amended on services and converters to allow for mocking and also to fix bugs the unit tests have found.

Example files have been added for testing under the Mocks folder.

A new configuration property has been added for SSL disable and also for port number altering.

Configuration property for Images have been set to not required.