Console Logging
Added a function to capture logging to the console before it actually writes it to check if it has gone through Log4Net. If it hasn't and doesn't match the Server Commands message then it gets put through Log4Net.

Also added the release date to all models which didn't previously have it

Will amend the Log4Net config through Chrome.